+++ Sources/StreamChatSwiftUI/ChatChannel/MessageList/MessageListConfig.swift
@@ -27,7 +27,8 @@ public struct MessageListConfig {
         showNewMessagesSeparator: Bool = true,
         handleTabBarVisibility: Bool = true,
         messageListAlignment: MessageListAlignment = .standard,
-        uniqueReactionsEnabled: Bool = false
+        uniqueReactionsEnabled: Bool = false,
+        localLinkDetectionEnabled: Bool = true
     ) {
         self.messageListType = messageListType
         self.typingIndicatorPlacement = typingIndicatorPlacement
@@ -48,6 +49,7 @@ public struct MessageListConfig {
         self.handleTabBarVisibility = handleTabBarVisibility
         self.messageListAlignment = messageListAlignment
         self.uniqueReactionsEnabled = uniqueReactionsEnabled
+        self.localLinkDetectionEnabled = localLinkDetectionEnabled
     }
 
     public let messageListType: MessageListType
@@ -69,6 +71,7 @@ public struct MessageListConfig {
     public let handleTabBarVisibility: Bool
     public let messageListAlignment: MessageListAlignment
     public let uniqueReactionsEnabled: Bool
+    public let localLinkDetectionEnabled: Bool
 }
 
 /// Contains information about the message paddings.